Add files via upload
The case of imprecision in sensing. There is a sensing error which gives us a normal pdf of observed sugar, with the true value as the mean. This is transformed to a pdf in the sensed expression level, and the mean of this is what the cells try to express, with some expression noise.